+++ gsn/src/gsn/wrappers/backlog/plugins/BinaryPlugin.java
@@ -178,7 +178,7 @@ public class BinaryPlugin extends AbstractPlugin {
 	
 
 	@Override
-	public byte getMessageType() {
+	public short getMessageType() {
 		return gsn.wrappers.backlog.BackLogMessage.BINARY_MESSAGE_TYPE;
 	}
